fastapi-app/requirements.txt
fastapi[standard]>=0.115.12
pydantic>=2.10.6
sqlalchemy>=2.0.40
uvicorn==0.34.0
psycopg2
bcrypt==4.0.1
passlib[bcrypt]
PyJWT
requests
redis
apscheduler